fixing errors of matrix caps

+++ tf2_models/em_routing.py
@@ -166,7 +166,7 @@ class EmRouting(tf.keras.layers.Layer):
     # Calculate kernel size by adding up column of spatial routing matrix
     # Do this before conventing the spatial_routing_matrix to tf
     tf.print(spatial_routing_matrix[:, 0].shape)
-    kk = tf.reduce_sum(spatial_routing_matrix[:, 0])
+    kk =tf.cast(tf.reduce_sum(spatial_routing_matrix[:, 0]), tf.int32)
 
     parent_caps = num_ouput_caps
     child_caps = int(kh_kw_i / kk)
